NEW:new helio ui/ux
