+++ resources/web/helio/helio_service_en.html
@@ -1,0 +1,109 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <style>
+        body {
+            position: absolute;
+            width: 700px;
+            height: 445px;
+            font-family: 'HarmonyOS Sans SC', sans-serif;
+            font-style: normal;
+            font-weight: 400;
+            font-size: 14px;
+            line-height: 22px; /* 157% */
+            color: #5C5C5C;
+            margin: 0;
+            box-sizing: border-box;
+        }
+        .header {
+            background-color: #f5f5f5;
+            padding: 10px 15px;
+            border-bottom: 1px solid #e0e0e0;
+        }
+		.underline {
+            text-decoration: underline;
+            text-underline-offset: 2px;
+        }
+        .header h1 {
+            font-size: 16px;
+            margin: 0;
+            color: #333333;
+        }
+        h2 {
+            font-size: 16px;
+            margin: 0 0 15px 0;
+            color: #333333;
+        }
+        p {
+            margin: 0 0 15px 0;
+            text-align: justify;
+        }
+        strong {
+            font-weight: 600;
+            color: #333333;
+        }
+        a {
+            color: #0066CC;
+            text-decoration: none;
+        }
+        a:hover {
+            text-decoration: underline;
+        }
+        ul {
+            margin: 0 0 15px 20px;
+            padding: 0;
+        }
+        li {
+            margin-bottom: 8px;
+        }
+        .buttons {
+            margin-top: 25px;
+            display: flex;
+            gap: 15px;
+        }
+        .btn {
+            padding: 8px 20px;
+            border-radius: 4px;
+            cursor: pointer;
+            font-family: 'HarmonyOS Sans SC', sans-serif;
+            font-size: 14px;
+            line-height: 1;
+            border: none;
+        }
+        .btn-primary {
+            background-color: #007DFF;
+            color: white;
+        }
+        .btn-primary:hover {
+            background-color: #0066CC;
+        }
+        .btn-secondary {
+            background-color: #F5F5F5;
+            color: #5C5C5C;
+            border: 1px solid #DDDDDD;
+        }
+        .btn-secondary:hover {
+            background-color: #EAEAEA;
+        }
+        .warning {
+            color: #FF5722;
+        }
+    </style>
+</head>
+<body>
+    <p>You are about to enable <strong>a third-party software service feature from <span class="underline">Helio Additive</span></strong>! Before confirming the use of this feature, please carefully read the following statements.</p>
+    
+    <p>Unless otherwise specified, Bambu Lab only provides support for the software features officially provided. The slicing evaluation and slicing optimization features based on Helio Additive's cloud service in this software will be <strong>developed, operated, provided, and maintained by Helio Additive. Helio Additive is responsible for the effectiveness and availability of this service</strong>. The optimization feature of this service may <strong>modify the default print commands, posing a risk of printer damage</strong>. These features will <strong>collect necessary user information and data</strong> to achieve relevant service functions. <strong>Subscriptions and payments may be involved</strong>. Please visit <a class="underline" href="#">Helio Additive</a> and refer to the <a class="underline" href="#">Helio Additive Privacy Agreement</a> and <a class="underline" href="#">Helio Additive User Agreement</a> for detailed information.</p>
+    
+    <p>Meanwhile, you understand that this product is provided to you "as is" <strong>based on <span class="underline">Helio Additive</span>'s services</strong>, and Bambu Lab makes no express or implied warranties of any kind, nor can it control the service effects. To the fullest extent permitted by applicable law, Bambu Lab or its licensors/affiliates do not provide any express or implied representations or warranties, including but not limited to warranties regarding merchantability, satisfactory quality, fitness for a particular purpose, accuracy, confidentiality, and non-infringement of third-party rights. Due to the nature of network services, Bambu Lab cannot guarantee that the service will be available at all times, and Bambu Lab reserves the right to terminate the service based on relevant circumstances. You agree not to use this product and its related updates to engage in the following activities:</p>
+    
+    <ol>
+        <li>Copy or use any part of this product outside the authorized scope of <span class="underline">Helio Additive</span> and Bambu Lab;</li>
+        <li>Attempt to disrupt, bypass, alter, invalidate, or evade any Digital Rights Management system related to and/or an integral part of this product;</li>
+        <li>Using this software and services for any improper or illegal activities.</li>
+    </ol>
+    
+    <p><strong>When you confirm to enable this feature, it means that you have confirmed and agreed to the above statements.</strong></p>
+</body>
+</html>
