+++ src/slic3r/GUI/Widgets/SwitchButton.hpp
@@ -95,6 +95,9 @@ public:
     void SetIsSelected(bool selected);
     bool IsSelected() const;
 
+    void set_primary_colour(wxColour col) {m_primary_colour = col;};
+    void set_secondary_colour(wxColour col) {m_secondary_colour = col;};
+
 private:
     void OnPaint(wxPaintEvent& event);
     void render(wxDC& dc);
@@ -106,6 +109,9 @@ private:
     wxString m_label;
     wxBitmap m_selected_icon;
     wxBitmap m_unselected_icon;
+    wxColour m_primary_colour{wxColour("#00AE42")};
+    wxColour m_secondary_colour{wxColour("#DEF5E7")};
+
     bool m_isSelected;
 };
